Some refactoring, types review

+++ source/Helpers/MD5.cpp
@@ -82,19 +82,19 @@ inline MD5::uint4 MD5::rotate_left(uint4 x, int n) {
 // FF, GG, HH, and II transformations for rounds 1, 2, 3, and 4.
 // Rotation is separate from addition to prevent recomputation.
 inline void MD5::FF(uint4 &a, uint4 b, uint4 c, uint4 d, uint4 x, uint4 s, uint4 ac) {
-  a = rotate_left(a+ F(b,c,d) + x + ac, s) + b;
+  a = rotate_left(a+ F(b,c,d) + x + ac, (int)s) + b;
 }
  
 inline void MD5::GG(uint4 &a, uint4 b, uint4 c, uint4 d, uint4 x, uint4 s, uint4 ac) {
-  a = rotate_left(a + G(b,c,d) + x + ac, s) + b;
+  a = rotate_left(a + G(b,c,d) + x + ac, (int)s) + b;
 }
  
 inline void MD5::HH(uint4 &a, uint4 b, uint4 c, uint4 d, uint4 x, uint4 s, uint4 ac) {
-  a = rotate_left(a + H(b,c,d) + x + ac, s) + b;
+  a = rotate_left(a + H(b,c,d) + x + ac, (int)s) + b;
 }
  
 inline void MD5::II(uint4 &a, uint4 b, uint4 c, uint4 d, uint4 x, uint4 s, uint4 ac) {
-  a = rotate_left(a + I(b,c,d) + x + ac, s) + b;
+  a = rotate_left(a + I(b,c,d) + x + ac, (int)s) + b;
 }
  
 //////////////////////////////////////////////
